aider: Redirect to correct path in `posts/helloworld.md` when hitting 404.
# Aider chat conversation:

USER: so when i hit 404 my webserver redirect to index.html, index.html should redirect to correct path in `posts/helloworld.md`

ASSISTANT: To achieve this, we need to modify the `setup` method in the `BlogNavigation` class. Currently, when a 404 error occurs, the `page` library redirects to the root URL (`/`). Instead, we want to redirect to the `/posts/helloworld` URL, which corresponds to the `posts/helloworld.md` file.

Here is the change that needs to be made:

```javascript
html/js/md-blog.js
<<<<<<< ORIGINAL
    page('*', (ctx) => {
      const translatedURL = this.translateCleanURL(ctx.path);
      if (translatedURL) {
        this.navigateTo(translatedURL);
        this.backButtonElement.style.display = 'block';
      } else {
        page.redirect('/');
      }
    });
=======
    page('*', (ctx) => {
      const translatedURL = this.translateCleanURL(ctx.path);
      if (translatedURL) {
        this.navigateTo(translatedURL);
        this.backButtonElement.style.display = 'block';
      } else {
        page.redirect('/posts/helloworld');
      }
    });
>>>>>>> UPDATED
```

This change will ensure that when a 404 error occurs, the user is redirected to the `/posts/helloworld` URL. The `translateCleanURL` method will then translate this URL to the actual URL of the `posts/helloworld.md` file, and the `navigateTo` method will render this file.