aider: Change `translateCleanURL` method to correctly translate URLs that include the filename within the `posts` directory.
# Aider chat conversation:

USER: here you hardcode helloworld but it just need to be the filename in posts/ directory for future possible blogposts

ASSISTANT: I see, you want to dynamically redirect to the correct blog post based on the URL. To achieve this, we need to modify the `translateCleanURL` method in the `BlogNavigation` class. This method is responsible for translating the clean URL to the actual URL of the markdown file.

Here is the change that needs to be made:

```javascript
html/js/md-blog.js
<<<<<<< ORIGINAL
  translateCleanURL(url) {
    const cleanURLRegex = /^\/posts\/(.+)$/;
    const match = url.match(cleanURLRegex);
    return match ? `${this.zeroMdUrl}${match[1]}` : null;
  }
=======
  translateCleanURL(url) {
    const cleanURLRegex = /^\/(.+)$/;
    const match = url.match(cleanURLRegex);
    return match ? `${this.zeroMdUrl}${match[1]}.md` : null;
  }
>>>>>>> UPDATED
```

This change will ensure that the `translateCleanURL` method correctly translates URLs that include the filename within the `posts` directory.