aider: Fix handleRender method to set the correct URL on the website when a link in a markdown file is clicked.
# Aider chat conversation:

USER: I ran this command:

 git push

And got this output:

To github.com:fjcloud/hello.git
   c876b3c..88af31c  main -> main

ASSISTANT: Great! The output indicates that your changes have been successfully pushed to the `main` branch of your repository on GitHub. The commit hash `88af31c` is now the latest commit on the `main` branch.

If you're still experiencing issues with the routing, please let me know so we can continue troubleshooting.

USER: when i click on a link of `index.md` it give me this path https://fjcloud.s3-website.fr-par.scw.cloud/content/fjcloud/hello/main/posts/helloworld.md, and it not work, i should have  https://fjcloud.s3-website.fr/posts/helloworld and should work!

ASSISTANT: The issue seems to be with the `handleRender` method in the `BlogNavigation` class. This method is responsible for handling the click event on the links in the markdown files. Currently, it's setting the `href` attribute of the link to the URL of the markdown file on GitHub, but it should be setting it to the URL on your website.

To fix this, we need to modify the `handleRender` method to set the `href` attribute of the link to the correct URL on your website.

Here is the change that needs to be made:

```javascript
html/js/md-blog.js
<<<<<<< ORIGINAL
  handleRender() {
    const nodes = this.zeroMdElement.shadowRoot.querySelectorAll('a[href$=".md"]');
    nodes.forEach((node) => {
      node.addEventListener('click', (e) => {
        e.preventDefault();
        const newSrc = node.href;
        this.navigateTo(newSrc);
        this.backButtonElement.style.display = 'block';
        const cleanHref = newSrc.match(/https:\/\/raw\.githubusercontent\.com(.+)/);
        if (cleanHref) {
          page(`/content${cleanHref[1]}`);
        }
      });
    });
  }
=======
  handleRender() {
    const nodes = this.zeroMdElement.shadowRoot.querySelectorAll('a[href$=".md"]');
    nodes.forEach((node) => {
      node.addEventListener('click', (e) => {
        e.preventDefault();
        const newSrc = node.href;
        this.navigateTo(newSrc);
        this.backButtonElement.style.display = 'block';
        const cleanHref = newSrc.match(/https:\/\/raw\.githubusercontent\.com\/fjcloud\/hello\/main\/(.+)/);
        if (cleanHref) {
          page(`/posts/${cleanHref[1]}`);
        }
      });
    });
  }
>>>>>>> UPDATED
```

This change will ensure that when a link in a markdown file is clicked, the user is redirected to the correct URL on your website.