aider: Fix issue with appending `.md` extension to the URL when a link in a markdown file is clicked.
# Aider chat conversation:

USER: when i click i have this url https://fjcloud.s3-website.fr-par.scw.cloud/posts/posts/helloworld.md i want this one https://fjcloud.s3-website.fr-par.scw.cloud/posts/helloworld

ASSISTANT: I see, the issue is that the `.md` extension is being included in the URL, but it should not be. This is happening because the `handleRender` method is appending `.md` to the URL.

To fix this, we need to modify the `handleRender` method to not append `.md` to the URL.

Here is the change that needs to be made:

```javascript
html/js/md-blog.js
<<<<<<< ORIGINAL
  handleRender() {
    const nodes = this.zeroMdElement.shadowRoot.querySelectorAll('a[href$=".md"]');
    nodes.forEach((node) => {
      node.addEventListener('click', (e) => {
        e.preventDefault();
        const newSrc = node.href.replace(/https:\/\/raw\.githubusercontent\.com\/fjcloud\/hello\/main\/(.+)/, '/posts/$1');
        this.navigateTo(newSrc);
        this.backButtonElement.style.display = 'block';
        page(newSrc);
      });
    });
  }
=======
  handleRender() {
    const nodes = this.zeroMdElement.shadowRoot.querySelectorAll('a[href$=".md"]');
    nodes.forEach((node) => {
      node.addEventListener('click', (e) => {
        e.preventDefault();
        const newSrc = node.href.replace(/https:\/\/raw\.githubusercontent\.com\/fjcloud\/hello\/main\/(.+)\.md/, '/posts/$1');
        this.navigateTo(newSrc);
        this.backButtonElement.style.display = 'block';
        page(newSrc);
      });
    });
  }
>>>>>>> UPDATED
```

This change will ensure that when a link in a markdown file is clicked, the user is redirected to the correct URL on your website without the `.md` extension.