aider: Fix issue with duplicated `/posts` in the generated URL by modifying the `handleRender` method to replace the correct part of the URL.
# Aider chat conversation:

USER: voila l'url généré ... https://fjcloud.s3-website.fr-par.scw.cloud/postsposts/helloworld ça devrait etre https://fjcloud.s3-website.fr-par.scw.cloud/posts/helloworld

ASSISTANT: I see, the issue is that the `/posts` part of the URL is being duplicated. This is happening because the `handleRender` method is appending `/posts` to the URL, but the URL already contains `/posts`.

To fix this, we need to modify the `handleRender` method to not append `/posts` to the URL.

Here is the change that needs to be made:

```javascript
html/js/md-blog.js
<<<<<<< ORIGINAL
  handleRender() {
    const nodes = this.zeroMdElement.shadowRoot.querySelectorAll('a[href$=".md"]');
    nodes.forEach((node) => {
      node.addEventListener('click', (e) => {
        e.preventDefault();
        const newSrc = node.href.replace(/https:\/\/raw\.githubusercontent\.com\/fjcloud\/hello\/main\/(.+)\.md/, '$1');
        this.navigateTo(newSrc);
        this.backButtonElement.style.display = 'block';
        page(newSrc);
      });
    });
  }
=======
  handleRender() {
    const nodes = this.zeroMdElement.shadowRoot.querySelectorAll('a[href$=".md"]');
    nodes.forEach((node) => {
      node.addEventListener('click', (e) => {
        e.preventDefault();
        const newSrc = node.href.replace(/https:\/\/raw\.githubusercontent\.com\/fjcloud\/hello\/main\/posts\/(.+)\.md/, '/posts/$1');
        this.navigateTo(newSrc);
        this.backButtonElement.style.display = 'block';
        page(newSrc);
      });
    });
  }
>>>>>>> UPDATED
```

This change will ensure that when a link in a markdown file is clicked, the user is redirected to the correct URL on your website without the `.md` extension and without duplicating the `/posts` part of the URL.